fox_note/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==4.2
django-allauth==64.0.0
gunicorn==23.0.0
idna==3.7
jwt==1.3.1
mysqlclient==2.2.4
packaging==24.1
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.2
